reboot: Add a simple plugin to reboot a machine
This introduces a new hook named 'wait_for_running' that should be
defined by the OS plugins to wait for the system to be ready. Since this
is the same for all systemd-based systems, group them in
os.common.red_hat_based and re-use them. While at it, move a couple of
other common hooks into this file as well.