backend.qemu: Use a pcie-root-port to attach devices
Upstream documentation [1] says that PCI and PCIe devices should not be
plugged directly into the root complex:

| Place only the following kinds of devices directly on the Root Complex:
|     (1) PCI Devices (e.g. network card, graphics card, IDE controller),
|         not controllers. Place only legacy PCI devices on
|         the Root Complex. These will be considered Integrated Endpoints.
|         Note: Integrated Endpoints are not hot-pluggable.
|
|         Although the PCI Express spec does not forbid PCI Express devices as
|         Integrated Endpoints, existing hardware mostly integrates legacy PCI
|         devices with the Root Complex. Guest OSes are suspected to behave
|         strangely when PCI Express devices are integrated
|         with the Root Complex.

So far, we have plugged additional disks directly into the root complex
(e.g., when attaching a swap disk). This seems to have worked fine on
x86_64, but it fails with qemu-system-aarch64 with the 'virt' machine.
Since the documentation quoted above discourages this use, migrate to
attaching a PCI Express Root Port to the root complex instead, and
attaching the disk to that port.

This works both on aarch64 and on x86_64, so there's no need to
distinguish the architectures here.

[1]: https://github.com/qemu/qemu/blob/master/docs/pcie.txt